REFACTOR(Application) : Game class를 분리
Game class를 분리하여 새로운 java 파일로 작성
